refactor: create reactions component
